\name{model.matrix.binaryChoice}
\alias{model.matrix.binaryChoice}

\title{Design Matrix of Binary Choice Models}

\description{
   Create design matrix of binary choice models
}

\usage{
\method{model.matrix}{binaryChoice}( object, ... )
}

\arguments{
   \item{object}{object of class \code{binaryChoice}.}
   \item{\dots}{currently not used.}
}

\value{
   The design matrix of binary choice models.
}

\author{Arne Henningsen,
   Ott Toomet \email{otoomet@ut.ee}}

\seealso{\code{\link{binaryChoice}}, \code{\link{probit}},
   \code{\link{model.matrix}}, \code{\link{model.frame.binaryChoice}},
   and \code{\link{probit-methods}}.}

\keyword{methods}
