\name{model.matrix.probit}
\alias{model.matrix.probit}

\title{Design Matrix of Probit Models}

\description{
   Create design matrix of probit models
}

\usage{
\method{model.matrix}{probit}( object, ... )
}

\arguments{
   \item{object}{object of class \code{probit}.}
   \item{\dots}{currently not used.}
}

\value{
   The design matrix of the probit model.
}

\author{Arne Henningsen \email{ahenningsen@agric-econ.uni-kiel.de},
   Ott Toomet \email{otoomet@ut.ee}}

\seealso{\code{\link{probit}},
   \code{\link{model.matrix}}, and \code{\link{model.frame.probit}}}

\keyword{methods}
