\name{sampleSelection-internal}
\alias{sampleSelection-internal}
\alias{tobitTfit}
\alias{lambda}
\alias{CB}
% Document the following:

%%%%


\title{ Internal sampleSelection Functions }

\description{
   Internal \code{sampleSelection} Functions
}

\details{
%   These are either various methods, or functions, not to be called
%   directly by the user (or in some cases are just
%   waiting for proper documentation to be written :).
   Currently, there are no internal functions in the sampleSelection package.
}

\keyword{ internal }
