\name{sampleSelection-internal}
\alias{coef.summary.probit}
\alias{model.frame.probit}
\alias{model.matrix.probit}
\alias{print.summary.probit}
\alias{summary.probit}
% Document the following:

%%%%


\title{ Internal sampleSelection Functions }

\description{
   Internal \code{sampleSelection} Functions
}

\details{
   These are either various methods, or functions, not to be called
   directly by the user (or in some cases are just
   waiting for proper documentation to be written :).
}

\keyword{ internal }
